integer : to.isotime, to.char and to.ascii fixed, mark strings as having or not substitutions

+++ source/now/commands/general.d
@@ -412,7 +412,7 @@ forLoop:
         SysTime today = Clock.currTime();
         long t = today.toUnixTime!long();
         log("- builtinCommands.unixtime.t:", t);
-        output.push(cast(int)t);
+        output.push(cast(long)t);
         return ExitCode.Success;
     };
